Moved logging enum to logging levels, syncronize datasets and sqlalchemy with the logging level
