+++ DashAI/back/dependencies/database/sqlite_database.py
@@ -1,9 +1,11 @@
+"""SQLite database module, implemented to be compatible with dependency injection."""
 import logging
 import pathlib
 from contextlib import contextmanager
 from typing import Callable, ContextManager
 
 from sqlalchemy import create_engine, orm
+from sqlalchemy.engine import Engine
 from sqlalchemy.orm import DeclarativeBase, Session
 
 logger = logging.getLogger(__name__)
@@ -14,7 +16,7 @@ class Base(DeclarativeBase):
 
 
 class SQLiteDatabase:
-    def __init__(self, db_path: pathlib.Path) -> None:
+    def __init__(self, db_path: pathlib.Path, logging_level: str) -> None:
         _db_path = str(db_path)
 
         if not _db_path.startswith("sqlite:///"):
@@ -22,9 +24,12 @@ class SQLiteDatabase:
 
         logger.info("Using %s as SQLite path.", db_url)
 
-        self._engine = create_engine(db_url, echo=True)
+        self._engine: Engine = create_engine(
+            db_url,
+            echo=logging_level == logging.DEBUG,
+        )
         self._session_factory = orm.scoped_session(
-            orm.sessionmaker(
+            session_factory=orm.sessionmaker(
                 autocommit=False,
                 autoflush=True,
                 bind=self._engine,
@@ -40,7 +45,6 @@ class SQLiteDatabase:
     @contextmanager
     def session(self) -> Callable[..., ContextManager[Session]]:
         session: Session = self._session_factory()
-
         try:
             logger.debug("Generating a new database session.")
             yield session
